Assignment_7/requirements.txt
streamlit
scikit-learn
numpy
matplotlib
seaborn
PyYAML
